Add test to ensure username is validated for uniqueness
